requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.11.17
distro==1.8.0
fire==0.5.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
human-eval==1.0
idna==3.6
multidict==6.0.4
numpy==1.26.2
openai==1.3.5
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
tenacity==8.2.3
termcolor==2.3.0
tqdm==4.66.1
typing_extensions==4.8.0
yarl==1.9.3